Refactor `cicd_utils/`
