+++ .github/workflows/ci.yaml
@@ -89,9 +89,9 @@ jobs:
         shell: bash
         run: tox -e tests-e2e
 
-      - name: Test the ci_utils
+      - name: Test the CI/CD utilities
         shell: bash
-        run: tox -e tests-ci_utils
+        run: tox -e tests-cicd_utils
 
       # Upload coverage reports to Codecov and Codacy.
