+++ cicd_utils/README.md
@@ -1,0 +1,18 @@
+# CI utilities
+
+This directory contains python source code and helper scripts used by CI/CD tasks.
+
+**Note to maintainers:** It is not mandatory, but it is highly recommended to also test the code in this directory (meta, I know!). See `tests/cicd_utils/` for some examples.
+
+## The `ci_pkg` package
+
+The `ci_pkg/` directory is a Python package containing modules used/imported by some CI tasks such as our test suite (`tests/`) or documentation building entrypoint (`docs/conf.py`). This package needs to be made explicitly discoverable to these tools by (for instance) appending it to the Python PATH:
+
+- For pytest, we configure this using the [pythonpath](https://docs.pytest.org/en/7.4.x/reference/reference.html#confval-pythonpath) option in `pytest.ini`.
+- For Sphinx, we set `PYTHONPATH={env:PYTHONPATH}{:}{toxinidir}/cicd_utils` in the relevant environment in `tox.ini`.
+- For mypy, we can add it to the `files` option in `mypy.ini` to help with import discovery.
+- For ruff/isort, we also added it to the `known-first-party` list (`ruff.toml`)
+
+## The `scripts` directory
+
+The `scripts` directory contains various scripts used by some CI/CD tasks. These scripts should all be marked as executable and contain a shebang line at the top.
