+++ cicd_utils/ci_pkg/ridgeplot_examples/_lincoln_weather.py
