+++ cicd_utils/scripts/extract_latest_release_notes.py
