+++ cicd_utils/scripts/open_in_browser.py
