+++ cicd_utils/scripts/update_colors_json.py
