+++ mypy.ini
@@ -2,7 +2,7 @@
 python_version = 3.8
 
 # Import discovery  ---
-files = src/ridgeplot, tests, docs, ci_utils
+files = src/ridgeplot, tests, docs, cicd_utils
 namespace_packages = False
 
 # Disallow dynamic typing  ---
