+++ pytest.ini
@@ -25,7 +25,7 @@ addopts =
 
 # List of directories that should be added to the python search path
 # https://docs.pytest.org/en/7.4.x/reference/reference.html#confval-pythonpath
-pythonpath = ci_utils
+pythonpath = cicd_utils
 
 # Tests marked with @pytest.mark.xfail that actually
 # succeed will by default fail the test suite.
