+++ requirements/docs.txt
@@ -18,6 +18,6 @@ sphinx_thebe
 sphinxext-opengraph
 # matplotlib is used to generate social cards
 matplotlib
-# ./ci_utils/_compile_plotly_charts.py
+# ./cicd_utils/ci_pkg/compile_plotly_charts.py
 minify-html
 kaleido
