+++ tests/e2e/test_examples.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Callable
 
 import pytest
 
-from _ridgeplot_examples import ALL_EXAMPLES
+from ci_pkg.ridgeplot_examples import ALL_EXAMPLES
 
 if TYPE_CHECKING:
     import plotly.graph_objects as go
