+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 labels =
     static = pre-commit, mypy
-    tests = tests-unit, tests-e2e, tests-ci_utils, doctests
+    tests = tests-unit, tests-e2e, tests-cicd_utils, doctests
 isolated_build = true
 
 [testenv]
@@ -25,7 +25,7 @@ deps =
     pytest-socket
     # e2e
     kaleido
-    # ci_utils
+    # cicd_utils
     beautifulsoup4
     minify-html
 commands =
@@ -33,7 +33,7 @@ commands =
     tests-unit: pytest tests/unit --cov={envsitepackagesdir}{/}ridgeplot {posargs:}
     tests-unit: diff-cover --compare-branch={env:DIFF_AGAINST:origin/main} coverage.xml
     tests-e2e: pytest tests/e2e {posargs:}
-    tests-ci_utils: pytest tests/ci_utils {posargs:}
+    tests-cicd_utils: pytest tests/cicd_utils {posargs:}
     doctests: pytest --doctest-modules src {posargs:}
 
 [testenv:pre-commit]
@@ -73,14 +73,14 @@ commands =
 description = generate Sphinx (live/static) HTML documentation
 setenv =
     {[testenv]setenv}
-    PYTHONPATH={env:PYTHONPATH}{:}{toxinidir}/ci_utils
+    PYTHONPATH={env:PYTHONPATH}{:}{toxinidir}/cicd_utils
 deps = -r requirements/docs.txt
 allowlist_externals = make
 commands =
     make --directory=docs clean
     live: make --directory=docs livehtml
     static: make --directory=docs html
-            python "ci_utils/scripts/open_in_browser.py" "docs/build/html/index.html"
+            python "cicd_utils/scripts/open_in_browser.py" "docs/build/html/index.html"
 
 [testenv:release-notes]
 description = extract release notes from docs/reference/changelog.md
@@ -88,7 +88,7 @@ skip_install = true
 deps =
     markdown-it-py
     mdformat
-commands = python ci_utils/scripts/extract_latest_release_notes.py
+commands = python cicd_utils/scripts/extract_latest_release_notes.py
 
 [testenv:publish-pypi-{test,prod}]
 description = build and upload the source and wheel package to (test/prod) PyPI
